requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
construct==2.5.3
Django==5.0.7
django-haystack==3.3.0
gunicorn==21.2.0
idna==3.7
packaging==24.1
pefile==2023.2.7
psycopg==3.2.1
psycopg2-binary==2.9.9
PyPDF2==3.0.1
python-dotenv==1.0.1
python-ptrace==0.9.9
requests==2.32.3
six==1.16.0
soupsieve==2.5
sqlparse==0.5.1
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
Whoosh==2.7.4
